+++ apps/common/include/openpfc_apps/moving_bc.hpp
@@ -100,7 +100,7 @@ public:
 
   void apply(Model &m, double time) override {
     (void)time;
-    const fft::IFFT &fft = get_fft(m);
+    const fft::IHostFFT &fft = get_fft(m);
     Field &field = get_real_field(m, get_field_name());
     const World &w = get_world(m);
     Int3 low = get_inbox(fft).low;